Modularized components and restructured files
