+++ components/buttons/BrewButton.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { useRouter } from 'next/router';
-// import PropTypes from 'prop-types';
 
 export default function BrewButton() {
   const router = useRouter();
