+++ components/cards/CreateNameCard.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import CreateNameModal from '../../modal/CreateNameModal';
+import { CreateNameModal } from '../modal';
 
 export default function CreateNameCard({ recipeObj, onUpdate }) {
   return (
