+++ components/cards/MethodEquipCard.js
