+++ components/icons/GuidedSteps.js
