+++ components/icons/index.js
@@ -1,0 +1,3 @@
+export { default as Grind } from './Grind';
+export { default as GuidedSteps } from './GuidedSteps';
+export { default as Method } from './Method';
