+++ components/index.js
@@ -1,0 +1,17 @@
+export {
+  SettingsModal, ChooseMethodModal, CreateNameModal, DeleteRecipeModal, EquipmentModal, MethodInfoModal, RatingsCommentsModal, StepModal, UserProcessModal, WaterWeightModal,
+} from './modal';
+export {
+  AddEquipButton, AddStepButton, AuthenticationButton, BrewButton,
+  EditDeleteEquip, EditDeleteStepsButtons,
+  FavoriteButton, PublishRecipeButton, ViewAllSteps,
+} from './buttons';
+export { Loading, Stopwatch } from './animated';
+export { default as NewUserForm } from './forms/NewUserForm';
+export { Grind, GuidedSteps, Method } from './icons';
+export { MainNavBar, UpperNav } from './navs';
+export {
+  ChooseBrewTime, ChooseGrindCard, ChooseMethodCard,
+  ChooseWeightCard, CreateNameCard, DefaultRecipes, EquipmentCard, FavRecipesCard, MethodEquipCard, ProfileCard, Recipes, StepCard,
+} from './cards';
+export { default as MethodCreate } from './MethodCreate';
