+++ components/modal/index.js
@@ -1,0 +1,11 @@
+export { default as SettingsModal } from './SettingsModal';
+export { default as ChooseBrewTimeModal } from './ChooseBrewTimeModal';
+export { default as ChooseMethodModal } from './ChooseMethodModal';
+export { default as CreateNameModal } from './CreateNameModal';
+export { default as DeleteRecipeModal } from './DeleteRecipeModal';
+export { default as EquipmentModal } from './EquipmentModal';
+export { default as MethodInfoModal } from './MethodInfoModal';
+export { default as RatingsCommentsModal } from './RatingsCommentsModal';
+export { default as StepModal } from './StepModal';
+export { default as UserProcessModal } from './UserProcessModal';
+export { default as WaterWeightModal } from './WaterWeightModal';
