+++ components/navs/index.js
@@ -1,0 +1,2 @@
+export { default as MainNavBar } from './MainNavBar';
+export { default as UpperNav } from './UpperNav';
