--- pages/read/grind/chooseGrind.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function ChooseGrind() {
-  return (
-    <div>chooseGrind</div>
-  );
-}
