--- pages/read/recipes/createName.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function CreateName() {
-  return (
-    <div>Create name</div>
-  );
-}
