+++ pages/read/user/createUser.js
@@ -5,7 +5,7 @@ import {
 } from 'react-bootstrap';
 import { useRouter } from 'next/router';
 import { useAuth } from '../../../utils/context/authContext';
-import NewUserForm from '../../../components/forms/NewUserForm';
+import { NewUserForm } from '../../../components';
 import { getUser } from '../../../utils/data/apiData/userData';
 import { checkUser } from '../../../utils/auth';
